requirements.txt
dj-database-url==0.4.1
Django==1.11.1
django-extensions==1.6.1
django-filter==1.0.4
django-material==1.0.0
django-model-utils==3.0.0
djangorestframework==3.6.3
gunicorn==19.6.0
ipython==6.2.1
psycopg2==2.7.3.1
python-env==1.0.0
whitenoise==3.2
